Ascii to debug angles as vertex property

+++ src/util/ply.h
@@ -40,7 +40,7 @@
 namespace colmap {
 
 struct PlyPointMetric {
-  uint8_t num_points = 0;
+  size_t num_points = 0;
   std::vector<float> normal_cosine_angles;
   std::vector<std::vector<float>> camera_camera_angles;
 };
